Templates new save full compatibility with old
AT conflicts solved by only updating the specific relationships
properties instead of the whole metadata

Using V1 mongoDatasource for updateLogs support

Trigger entities update events on each entity when denormalizing.
